# specs/genesys-openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "With the PureCloud Platform API, you can control all aspects of your PureCloud environment. With the APIs you can access the system configuration, manage conversations and more.",
    "version": "v2",
    "title": "PureCloud Platform API",
    "termsOfService": "https://help.genesys.cloud/articles/terms-and-conditions/",
    "contact": {
      "name": "PureCloud Developer Evangelists",
      "url": "https://developer.genesys.cloud/",
      "email": "DeveloperEvangelists@genesys.com"
    },
    "license": {
      "name": "UNLICENSED",
      "url": "https://help.genesys.cloud/articles/terms-and-conditions/"
    }
  },
  "host": "api.mypurecloud.com",
  "basePath": null,
  "paths": {
...
    "/api/v2/authorization/divisions": {
      "get": {
        "operationId": "getAuthorizationDivisions",
        "summary": "Retrieve a list of all divisions defined for the organization",
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "schema": null
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "type": "integer",
            "schema": null
          }
        ]
      }
    },
...
    "/api/v2/routing/users/{userId}/utilization": {
      "get": {
        "operationId": "getRoutingUserUtilization",
        "summary": "Get the user's max utilization settings",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          }
        ]
      },
      "put": {
        "operationId": "putRoutingUserUtilization",
        "summary": "Update the user's max utilization settings",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          }
        ]
      },
      "delete": {
        "operationId": "deleteRoutingUserUtilization",
        "summary": "Delete the user's max utilization settings; revert to org default",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          }
        ]
      }
    },
    "/api/v2/routing/utilization": {
      "get": {
        "operationId": "getRoutingUtilization",
        "summary": "Get the organization-wide max utilization settings.",
        "parameters": []
      },
      "put": {
        "operationId": "putRoutingUtilization",
        "summary": "Update the organization-wide max utilization settings.  Include only those media types requiring custom configuration. Any omitted media types will be set to the system default values.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": null,
            "schema": {
              "$ref": "#/definitions/UtilizationRequest"
            }
          }
        ]
      },
      "delete": {
        "operationId": "deleteRoutingUtilization",
        "summary": "Delete the organization-wide max utilization settings and revert to the system default.",
        "parameters": []
      }
    },
...
    "/api/v2/tokens/me": {
      "get": {
        "operationId": "getTokensMe",
        "summary": "Fetch information about the current token",
        "parameters": [
          {
            "name": "preserveIdleTTL",
            "in": "query",
            "required": false,
            "type": "boolean",
            "schema": null
          }
        ]
      },
      "head": {
        "operationId": "headTokensMe",
        "summary": "Verify user token",
        "parameters": []
      },
      "delete": {
        "operationId": "deleteTokensMe",
        "summary": "Delete auth token used to make the request.",
        "parameters": []
      }
    },
...
  }
}